feat(core): process() entry point + World + Event/Effect + 17 integration tests
Completes the pure-core surface area per SPEC §5.5. A single pure
function:

    process(event, world, clock, topology) -> Vec<Effect>

applies the event to the world (sensor reading, typed-sensor reading,
readback, command, or Tick), runs all six controllers, and returns the
resulting effects.

Added:

- crates/core/src/types.rs — closed enums for every Event, Effect, ID:
    * SensorId (20 scalar sensors), ActuatedId (6), KnobId (26),
      ForecastProvider (3), ScheduleField (5).
    * KnobValue: typed variants (Bool / Float / Uint32 / DischargeTime /
      DebugFullCharge / ForecastDisagreementStrategy).
    * Event: Sensor / TypedSensor / Readback / Command / Tick.
    * Command: Knob | KillSwitch. Owner carried at the envelope level.
    * Effect: WriteDbus / CallMyenergi / Publish / Log.
    * PublishPayload: Knob | ActuatedPhase | KillSwitch | Bookkeeping.

- crates/core/src/topology.rs — ControllerParams (tolerances + freshness
  thresholds per SPEC §5.3 defaults) and Topology wrapper.

- crates/core/src/world.rs — top-level World struct with:
    * 20 scalar sensor Actuals + 2 typed (Zappi/Eddi) + 3 forecast
      snapshots.
    * 6 actuated entities.
    * Knobs (all fields f64 now — matches controller signatures and JS
      number semantics).
    * Bookkeeping (cross-controller state: next_full_charge, above_soc_date,
      prev_ess_state, zappi_active, charge_to_full_required, …).
    * KnobProvenance for the γ-hold window.
    * Fresh-boot constructor installing safe_defaults + unknown sensors.

- crates/core/src/process.rs — the process() function + controller
  dispatchers. Each of the 6 controllers (setpoint, current_limit,
  schedules, zappi_mode, eddi_mode, weather_soc) is wired through:
    * Inputs assembled from World/Knobs/Bookkeeping.
    * Required-sensor freshness gate → `None` / safety fallback
      if anything critical is Stale/Unknown.
    * Proposal of a target (with dead-band filter for scalar actuated
      entities).
    * If writes_enabled=false: the proposal updates the TASS phase but
      no WriteDbus / CallMyenergi effect is emitted (observer mode).
    * Otherwise: emit effect + mark_commanded.
    * Bookkeeping updates published when they actually change.
  Safety policies per SPEC §5.13:
    * Setpoint: Battery SoC stale/unknown → force 10 W, owner=System.
    * Eddi: SoC not Fresh → force Stopped, bypassing dwell.
    * γ-rule: HA commands within 1 s of a Dashboard write are dropped
      with a Log effect; accepted after the window.

17 integration tests in process::tests cover the full flow:

- setpoint_proposes_and_commands_when_all_sensors_fresh
- setpoint_freezes_at_10w_when_battery_soc_stale (safety)
- setpoint_is_not_emitted_when_writes_enabled_false (kill switch)
- setpoint_deadband_suppresses_minor_changes
- setpoint_phase_advances_to_confirmed_on_matching_readback
- setpoint_retargets_on_consumption_change_beyond_deadband
- ha_knob_command_suppressed_within_dashboard_hold (γ rule)
- ha_knob_command_accepted_after_hold_expires
- kill_switch_toggles_writes_enabled_and_publishes
- eddi_requires_fresh_soc
- eddi_sets_normal_when_soc_above_threshold_and_current_known_stopped
- eddi_safety_stops_when_soc_becomes_stale
- schedules_emit_five_writes_when_target_changes
- second_tick_with_no_change_does_not_re_emit_schedule_writes
- tick_decays_sensor_freshness_fresh_to_stale
- sensor_event_updates_world_field
- typed_sensor_event_updates_zappi_state

Knobs struct converted from f32 to f64 throughout for uniformity with
controllers; arithmetic matches JS number semantics more cleanly.

171 tests total pass; clippy --all-targets -- -D warnings clean.

Known shortcut: the weather_soc planner writes `charge_battery_extended`
as a folded bit of `charge_to_full_required` rather than the legacy
separately-tracked global. This keeps the integration simple for now;
will split back out when forecast wiring lands in M5.

+++ crates/core/src/knobs.rs
@@ -46,26 +46,26 @@ pub struct Knobs {
     // --- Export / discharge policy ---
     pub force_disable_export: bool,
     /// Export when battery SoC (%) is at or above this threshold.
-    pub export_soc_threshold: f32,
+    pub export_soc_threshold: f64,
     /// Evening controller targets this SoC (%) at end-of-day.
-    pub discharge_soc_target: f32,
+    pub discharge_soc_target: f64,
     /// Night-time scheduled charge target (%).
-    pub battery_soc_target: f32,
+    pub battery_soc_target: f64,
     /// Evening target (%) during weekly full-charge.
-    pub full_charge_discharge_soc_target: f32,
+    pub full_charge_discharge_soc_target: f64,
     /// Export threshold (%) during weekly full-charge.
-    pub full_charge_export_soc_threshold: f32,
+    pub full_charge_export_soc_threshold: f64,
     pub discharge_time: DischargeTime,
     pub debug_full_charge: DebugFullCharge,
-    pub pessimism_multiplier_modifier: f32,
+    pub pessimism_multiplier_modifier: f64,
     pub disable_night_grid_discharge: bool,
 
     // --- Zappi ---
     pub charge_car_boost: bool,
     pub charge_car_extended: bool,
-    pub zappi_current_target: f32,
-    pub zappi_limit: f32,
-    pub zappi_emergency_margin: f32,
+    pub zappi_current_target: f64,
+    pub zappi_limit: f64,
+    pub zappi_emergency_margin: f64,
 
     // --- New knobs (SPEC §2.10a) ---
     /// Hard cap on negative setpoint magnitude (grid-side export limit, W).
@@ -74,18 +74,18 @@ pub struct Knobs {
     /// windows. Always boots `false` regardless of retained value.
     pub allow_battery_to_car: bool,
     /// Eddi target becomes Normal when battery SoC ≥ this (%).
-    pub eddi_enable_soc: f32,
+    pub eddi_enable_soc: f64,
     /// Eddi target becomes Stopped when battery SoC ≤ this (%).
-    pub eddi_disable_soc: f32,
+    pub eddi_disable_soc: f64,
     /// Minimum dwell time (s) at the current Eddi state before re-evaluation.
     pub eddi_dwell_s: u32,
 
     // --- Weather-SoC planner thresholds ---
-    pub weathersoc_winter_temperature_threshold: f32,
-    pub weathersoc_low_energy_threshold: f32,
-    pub weathersoc_ok_energy_threshold: f32,
-    pub weathersoc_high_energy_threshold: f32,
-    pub weathersoc_too_much_energy_threshold: f32,
+    pub weathersoc_winter_temperature_threshold: f64,
+    pub weathersoc_low_energy_threshold: f64,
+    pub weathersoc_ok_energy_threshold: f64,
+    pub weathersoc_high_energy_threshold: f64,
+    pub weathersoc_too_much_energy_threshold: f64,
 
     // --- Ops ---
     pub writes_enabled: bool,
@@ -145,17 +145,17 @@ mod tests {
         let k = Knobs::safe_defaults();
         // Spot-check the non-trivial values from SPEC §7.
         assert!(!k.force_disable_export);
-        assert!((k.export_soc_threshold - 80.0).abs() < f32::EPSILON);
-        assert!((k.discharge_soc_target - 80.0).abs() < f32::EPSILON);
-        assert!((k.battery_soc_target - 80.0).abs() < f32::EPSILON);
-        assert!((k.full_charge_discharge_soc_target - 57.0).abs() < f32::EPSILON);
-        assert!((k.full_charge_export_soc_threshold - 100.0).abs() < f32::EPSILON);
+        assert!((k.export_soc_threshold - 80.0).abs() < f64::EPSILON);
+        assert!((k.discharge_soc_target - 80.0).abs() < f64::EPSILON);
+        assert!((k.battery_soc_target - 80.0).abs() < f64::EPSILON);
+        assert!((k.full_charge_discharge_soc_target - 57.0).abs() < f64::EPSILON);
+        assert!((k.full_charge_export_soc_threshold - 100.0).abs() < f64::EPSILON);
         assert_eq!(k.discharge_time, DischargeTime::At0200);
         assert_eq!(k.debug_full_charge, DebugFullCharge::None);
         assert_eq!(k.grid_export_limit_w, 4900);
         assert!(!k.allow_battery_to_car);
-        assert!((k.eddi_enable_soc - 96.0).abs() < f32::EPSILON);
-        assert!((k.eddi_disable_soc - 94.0).abs() < f32::EPSILON);
+        assert!((k.eddi_enable_soc - 96.0).abs() < f64::EPSILON);
+        assert!((k.eddi_disable_soc - 94.0).abs() < f64::EPSILON);
         assert_eq!(k.eddi_dwell_s, 60);
         assert!(k.writes_enabled);
     }
